Initial edit before I dive headfirst into UndoRedo
Added some QoL that clears up some confusions found on the TC discord. (Added a couple of tooltips, and changed the default .tmb to be in 4/4 with an endpoint of 4)

For some reason, Godot is taking away all the button offsets of every file that I access, and changing "Vector2i"'s to "Vector2"'s.

Godot also annihilated project.godot upon loading the repo. I don't know if that data is unnecessary on the most recent version, or what.

If any of the above breaks anything, let me know, and I will manually deselect those changes next PR.